+++ tinydl/trainer.py
@@ -2,8 +2,6 @@ from tinydl.layers import *
 from tinydl.helpers import *
 from tinydl.config import *
 from tinydl.loss import *
-from tinydl.logger import *
-from tinydl.autograd import *
 import matplotlib.pyplot as plt
 
 if usegpu == True:
@@ -13,82 +11,15 @@ else:
 # list of activation functions except sigmoid and tanh
 act_layers = ["softplus", "relu", "prelu", "lrelu", "elu", "swish"]
 
-def dropout(arr, p = 0.5):
-    return arr* np.random.binomial(1, p, size = arr.shape)
-
-def linearInit(arch, param_values, seed=99):
-    # for linear layers
-    np.random.seed(seed)
-
-    for idx, layer in enumerate(arch):
-        current_idx = idx+1
-        if layer.name in ["linear"]:
-            layer_in_size = layer["input_dim"]
-            layer_out_size = layer["output_dim"]
-
-            param_values['W' + str(current_idx)] = Tensor(np.random.randn(
-                layer_out_size, layer_in_size
-            ) * 0.1)
-            param_values['b' + str(current_idx)] = Tensor(np.random.randn(
-                layer_out_size, 1
-            ) * 0.1)
-
-    return param_values
-
-
-def xavierInit(arch, param_values, seed=99):
-    # for activation functions, mostly for tanh
-    np.random.seed(seed)
-    for idx, layer in enumerate(arch):
-        current_idx = idx+1
-        if layer["name"] in ["tanh", "sigmoid"]:
-            n = layer["input_dim"]
-            lower, upper = -(1.0 / np.sqrt(n)), (1.0/np.sqrt(n))
-            nums = np.random.randn(1000)
-            param_values['W' + str(current_idx)] = Tensor(lower + nums*(upper - lower))
-    return param_values
-
-
-def normalizedXavierInit(arch, param_values, seed=99):
-    # for activation functions - normalized Xavier
-    np.random.seed(seed)
-    for idx, layer in enumerate(arch):
-        current_idx = idx+1
-        if layer["name"] in ["tanh", "sigmoid"]:
-            n = layer["input_dim"]
-            m = layer["output_dim"]
-            lower, upper = -(np.sqrt(6.0) / np.sqrt(n + m)
-                             ), (np.sqrt(6.0) / np.sqrt(n + m))
-            nums = np.random.randn(1000)
-            param_values['W' + str(current_idx)] = Tensor(lower + nums*(upper - lower))
-    return param_values
-
-
-def heInit(arch, param_values, seed=99):
-    # for activation functions, mostly for relu
-    np.random.seed(seed)
-    for idx, layer in enumerate(arch):
-        current_idx = idx+1
-        if layer["name"] in act_layers:
-            n = layer["input_dim"]
-            lower, upper = -(1.0 / np.sqrt(n)), (1.0/np.sqrt(n))
-            nums = np.random.randn(1000)
-            param_values['W' + str(current_idx)] = Tensor(lower + nums*(upper - lower))
-    return param_values
-
-
-def defaultInit(arch, seed=99):
-    # basically just use linear for linear layers and He init for activations
-    param_values = {}
-    linearInit(arch, param_values=param_values)
-    heInit(arch, param_values=param_values)
-    normalizedXavierInit(arch, param_values=param_values)
-    return param_values
-
+# optimizer
 def SGD(model,k):
     new_lr = 1.0 - 0.9*k/100
     for p in model.parameters():
         p.data -= new_lr * p.grad
+
+def GD(model,lr=lr):
+    for p in model.parameters():
+        p.data -= lr * p.grad
 
 def ADAM(param_values, gradsVals, arch, lr=0.01):
     pass
@@ -96,6 +27,7 @@ def ADAM(param_values, gradsVals, arch, lr=0.01):
 dict_optim = {
     "ADAM" : ADAM,
     "SGD" : SGD,
+    "GD" : GD,
 }
 
 dict_loss = {
@@ -113,14 +45,12 @@ def forward(x, y, model, bs = None):
         xb, yb = x[rinde], y[rinde]
 
     inputs = [list(map(Tensor, xrow)) for xrow in xb]
+    if layerdropout == True:
+        for layer in model.layers:
+            dropout(layer.parameters(), layerdropoutprob)
     out = list(map(model, inputs))
 
-    losses = [(1 + -yi*scorei).relu() for yi, scorei in zip(yb, out)]
-    data_loss = sum(losses) * (1.0 / len(losses))
-    # L2 regularization
-    alpha = 1e-4
-    reg_loss = alpha * sum((p*p for p in model.parameters()))
-    total_loss = data_loss + reg_loss
+    total_loss = dict_loss[lossfunc](model, yb, out)
     
     acc = [(yi > 0) == (scorei.data > 0) for yi, scorei in zip(yb, out)]
     total_acc = sum(acc) / len(acc)
@@ -159,6 +89,4 @@ def train(x,y, model, epochs = 1, lr= 0.001,bs = batchsize, verbose = True, afte
     if plotAcc == True:
         plt.plot(acchistory)
     plt.show()
-
-
